% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoloGEx.R
\name{check_genes_match}
\alias{check_genes_match}
\alias{check_genes_match,SoloGEx,SoloGEx-method}
\title{Check gene consistency between two SoloGEx objects}
\usage{
check_genes_match(obj1, obj2)

\S4method{check_genes_match}{SoloGEx,SoloGEx}(obj1, obj2)
}
\arguments{
\item{obj1}{SoloGEx object (e.g., singletons)}

\item{obj2}{SoloGEx object (e.g., larger reference set)}
}
\value{
TRUE if genes match, stops with an informative error otherwise
}
\description{
Check gene consistency between two SoloGEx objects
}
\section{Methods (by class)}{
\itemize{
\item \code{check_genes_match(obj1 = SoloGEx, obj2 = SoloGEx)}: Method for SoloGEx objects

}}
