% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoloGEx.R
\name{compute_corr_to_reference}
\alias{compute_corr_to_reference}
\alias{compute_corr_to_reference,SoloGEx,SoloGEx-method}
\title{Compute correlation of singleton samples to reference samples (S4 method)}
\usage{
compute_corr_to_reference(singleton, reference, method = "pearson")

\S4method{compute_corr_to_reference}{SoloGEx,SoloGEx}(singleton, reference, method = "pearson")
}
\arguments{
\item{singleton}{SoloGEx object containing singleton expression data}

\item{reference}{SoloGEx object containing reference expression data}

\item{method}{Correlation method: "pearson" or "spearman" (default: "pearson")}
}
\value{
SoloGEx object with correlations stored in @singleton_analysis$correlation_to_reference
}
\description{
Calculates correlation between each singleton sample and all reference samples.
}
\section{Methods (by class)}{
\itemize{
\item \code{compute_corr_to_reference(singleton = SoloGEx, reference = SoloGEx)}: Method for SoloGEx objects

}}
