% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoloGEx.R
\docType{class}
\name{SoloGEx-class}
\alias{SoloGEx-class}
\title{SoloGEx constructor}
\arguments{
\item{expr}{numeric matrix of expression values (genes x samples)}

\item{colData}{data.frame with sample metadata (rows = samples)}
}
\value{
SoloGEx S4 object
}
\description{
SoloGEx constructor
}
