% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoloGEx.R
\name{enforce_gene_order}
\alias{enforce_gene_order}
\title{Reorder genes of target SoloGEx to match reference SoloGEx}
\usage{
enforce_gene_order(target, reference)
}
\arguments{
\item{target}{SoloGEx object to reorder}

\item{reference}{SoloGEx object whose gene order will be used}
}
\value{
target SoloGEx object with rows reordered to match reference
}
\description{
Reorder genes of target SoloGEx to match reference SoloGEx
}
